feat: add higher order feeding comm reducer
The higher order feeding communication reducer
better encapsulates the communication with the parent
app!

Eventually (and as noted in comment in code), the
parent app communication and local feeding state
management should probably be flipped in order to better
deal with possible failure in communication.
